git-svn-id: https://svn.r-project.org/R/branches/ALTREP@71676 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/sysutils.c
@@ -882,7 +882,8 @@ next_char:
 # ifndef Win32
 		if((unsigned int) wc < 65536) {
 # endif
-		    snprintf(outbuf, 9, "<U+%04X>", (unsigned int) wc);
+		// gcc 7 objects to this with unsigned int
+		    snprintf(outbuf, 9, "<U+%04X>", (unsigned short) wc);
 		    outbuf += 8; outb -= 8;
 # ifndef Win32
 		} else {
